Updating the test configuration
